requirements.txt
torch
torchvision
pytorch_lightning
pillow
pandas
# wandb
